Separate concerns in fixups.
