+++ vscode/src/jsonrpc/agent-protocol.ts
@@ -329,9 +329,9 @@ interface ClientCapabilities {
     showDocument?: 'none' | 'enabled'
     codeLenses?: 'none' | 'enabled'
     showWindowMessage?: 'notification' | 'request'
-    // The strategy for decorating fixups with controls in code lenses.
-    // To enable, must also have codeLenses == 'enabled'.
-    fixupCodeLenses?: 'none' | 'enabled'
+    // How to render fixup controls.
+    // To use 'lenses', must also have codeLenses === 'enabled'.
+    fixupControls?: 'none' | 'lenses'
 }
 
 export interface ServerInfo {
